requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.11.1
Django==5.0.6
django-jazzmin==3.0.0
django-resized==1.0.2
djangorestframework==3.14.0
idna==3.7
imagesize==1.3.0
lxml==5.2.2
more-itertools==10.3.0
pillow==10.3.0
premailer==3.10.0
pytz==2024.1
requests==2.32.3
sorl-thumbnail==12.10.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
yagmail==0.15.293
